solution to 13 - Hangman - 13
