+++ .gitignore
@@ -1,1 +1,3 @@
 .DS_Store
+__pycache__/
+*.pyc
